# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ShopXperience API",
    "description": "Multi-tenant Point of Sale System API",
    "version": "1.0.0",
    "contact": {
      "name": "ShopXperience Support",
      "email": "support@shopxperience.com"
    }
  },
  "servers": [
    {
      "url": "https://api.shopxperience.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "sku": { "type": "string" },
          "price": { "type": "number", "format": "decimal" },
          "cost": { "type": "number", "format": "decimal" },
          "category": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Transaction": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "customer_id": { "type": "integer" },
          "location_id": { "type": "integer" },
          "total_amount": { "type": "number", "format": "decimal" },
          "tax_amount": { "type": "number", "format": "decimal" },
          "discount_amount": { "type": "number", "format": "decimal" },
          "payment_method": { "type": "string" },
          "status": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "summary": "Get all products",
        "description": "Retrieve paginated list of products with optional filters",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "page": { "type": "integer" },
                        "pages": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "summary": "Create product",
        "description": "Create a new product (admin/manager only)",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "sku", "price"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "sku": { "type": "string" },
                  "price": { "type": "number", "format": "decimal" },
                  "cost": { "type": "number", "format": "decimal" },
                  "category": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "product": { "$ref": "#/components/schemas/Product" }
                  }
                }
              }
            }
          },
          "409": { "$ref": "#/components/responses/ConflictError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Get product by ID",
        "description": "Retrieve a specific product",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Product retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product": { "$ref": "#/components/schemas/Product" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "summary": "Update product",
        "description": "Update product information (admin/manager only)",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "sku": { "type": "string" },
                  "price": { "type": "number", "format": "decimal" },
                  "cost": { "type": "number", "format": "decimal" },
                  "category": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "product": { "$ref": "#/components/schemas/Product" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "409": { "$ref": "#/components/responses/ConflictError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete product",
        "description": "Delete a product (admin/manager only)",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequestError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/transactions": {
      "get": {
        "summary": "Get all transactions",
        "description": "Retrieve paginated list of transactions with optional filters",
        "tags": ["Transactions"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "location_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "customer_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "start_date", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "end_date", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "Transactions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transactions": { "type": "array", "items": { "$ref": "#/components/schemas/Transaction" } },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "page": { "type": "integer" },
                        "pages": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "summary": "Create transaction",
        "description": "Create a new transaction (cash payment only)",
        "tags": ["Transactions"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["location_id", "items", "payment_method"],
                "properties": {
                  "customer_id": { "type": "integer" },
                  "location_id": { "type": "integer" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["product_id", "quantity"],
                      "properties": {
                        "product_id": { "type": "integer" },
                        "quantity": { "type": "integer" }
                      }
                    }
                  },
                  "payment_method": { "type": "string", "enum": ["cash", "check", "other"] },
                  "discount_amount": { "type": "number", "format": "decimal", "default": 0 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "transaction": { "$ref": "#/components/schemas/Transaction" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequestError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/transactions/{id}": {
      "get": {
        "summary": "Get transaction by ID",
        "description": "Retrieve a specific transaction with items",
        "tags": ["Transactions"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Transaction retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transaction": { "$ref": "#/components/schemas/Transaction" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequestError": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "NotFoundError": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "ConflictError": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  }
}